requirements.txt
Adafruit-Blinka==6.0.1
adafruit-circuitpython-bme680==3.3.2
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-dht==3.5.4
Adafruit-PlatformDetect==3.0.0
Adafruit-PureIO==1.1.8
aiounittest==1.4.0
appdirs==1.4.4
asyncio==3.4.3
attrs==20.3.0
black==20.8b1
boto3==1.17.1
boto3-stubs==1.17.47.0
botocore==1.20.1
certifi==2020.12.5
click==7.1.2
flake8==3.8.4
iniconfig==1.1.1
jmespath==0.10.0
mccabe==0.6.1
mypy==0.800
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyftdi==0.52.9
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.2
pytest-asyncio==0.15.0
python-dateutil==2.8.1
pyusb==1.1.1
regex==2020.11.13
RPi.GPIO-def==0.2
s3transfer==0.3.4
sentry-sdk==1.0.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
w1thermsensor==2.0.0
websockets==8.1
wrapt==1.12.1
